+++ media/smart-card/src/constants.ts
@@ -76,7 +76,7 @@ export enum SmartLinkWidth {
 /**
  * Flexible UI element name - each reflecting the link data its represented.
  * When adding an element...
- * 1) Create base element if it doesn't already existed.
+ * 1) Create base element if it doesn't already exist.
  *    Base element are inside src/view/FlexibleCard/components/elements.
  *    E.g. Badge,DateTime, Icon, Lozenge, etc.
  * 2) Update FlexibleUiContext with the new prop for data representing
@@ -109,6 +109,7 @@ export enum ElementName {
   Title = 'Title',
   ViewCount = 'ViewCount',
   VoteCount = 'VoteCount',
+  LatestCommit = 'LatestCommit',
 }
 
 /**