+++ media/smart-card/src/view/FlexibleCard/components/blocks/utils.tsx
@@ -40,6 +40,7 @@ export const ElementDisplaySchema: Record<
   [ElementName.SubscriberCount]: ['inline'],
   [ElementName.Title]: ['inline'],
   [ElementName.Provider]: ['inline'],
+  [ElementName.LatestCommit]: ['inline'],
 };
 
 const getDirectionStyles = (