+++ editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atlaskit/editor-mobile-bridge
 
+## 34.2.11
+
+### Patch Changes
+
+- Updated dependencies
+
+## 34.2.10
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 34.2.9
 
 ### Patch Changes