+++ linking-platform/linking-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/linking-common
 
+## 1.1.1
+
+### Patch Changes
+
+- [`50b81e07a35`](https://bitbucket.org/atlassian/atlassian-frontend/commits/50b81e07a35) - Version of package 'json-ld-types' was upgraded to 2.4.2
+
 ## 1.1.0
 
 ### Minor Changes